scraper/requirements.txt
# FastAPI y servidor
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5

# Crawl4AI para scraping inteligente
crawl4ai[all]==0.7.8

# HTTP client for LLM API calls
httpx==0.28.1

# Utilidades
python-dotenv==1.2.1
